Merge dev branch (#38)
* Update: restructure code

separate the code logic from the sql statements to add
a distinct layer between executing statements against
the database and the success and failure of the action,
so logging and errors are processed on the new layer
and the logic is easier to follow and debug.

* Added: retry mechanism to process failed jobs

create_player function now adds failed attempts to failed[]
under job type 1 for further attempts. Failure eventually
kicks the player. Possibility now exists to code other db
insert and update functions as jobs if they fail after first
attempt

* Added: ban status text field to the ban gui

* Update: comments & log event